move apply grid loss to grid loss aggregator
