+++ source/databricks/calculation_engine/package/calculation/energy/aggregators/grid_loss_aggregators.py
@@ -20,6 +20,7 @@ import pyspark.sql.functions as f
 import package.calculation.energy.aggregators.transformations as t
 from package.constants import Colname
 from package.calculation.energy.energy_results import EnergyResults
+from package.calculation.preparation.grid_loss_responsible import GridLossResponsible
 
 production_sum_quantity = "production_sum_quantity"
 exchange_sum_quantity = "exchange_sum_quantity"
@@ -159,6 +160,108 @@ def calculate_total_consumption(
         Colname.qualities,
         Colname.sum_quantity,
         f.lit(MeteringPointType.CONSUMPTION.value).alias(Colname.metering_point_type),
+    )
+
+    return EnergyResults(result)
+
+
+def apply_grid_loss_adjustment(
+    results: EnergyResults,
+    grid_loss_result: EnergyResults,
+    grid_loss_responsible: GridLossResponsible,
+    metering_point_type: MeteringPointType,
+) -> EnergyResults:
+    grid_loss_responsible_energy_supplier = "GridLossResponsible_EnergySupplier"
+    grid_loss_responsible_grid_area = "GridLossResponsible_GridArea"
+    adjusted_sum_quantity = "adjusted_sum_quantity"
+
+    result_df = results.df
+    grid_loss_result_df = grid_loss_result.df
+
+    grid_loss_responsible_df = grid_loss_responsible.df.select(
+        Colname.from_date,
+        Colname.to_date,
+        f.col(Colname.energy_supplier_id).alias(grid_loss_responsible_energy_supplier),
+        f.col(Colname.grid_area).alias(grid_loss_responsible_grid_area),
+        Colname.metering_point_type,
+    )
+
+    joined_grid_loss_result_and_responsible = grid_loss_result_df.join(
+        grid_loss_responsible_df,
+        f.when(
+            f.col(Colname.to_date).isNotNull(),
+            f.col(Colname.time_window_start) <= f.col(Colname.to_date),
+        ).otherwise(True)
+        & (f.col(Colname.time_window_start) >= f.col(Colname.from_date))
+        & (
+            f.col(Colname.to_date).isNull()
+            | (f.col(Colname.time_window_end) <= f.col(Colname.to_date))
+        )
+        & (f.col(Colname.grid_area) == f.col(grid_loss_responsible_grid_area))
+        & (f.col(Colname.metering_point_type) == metering_point_type.value),
+        "left",
+    ).select(
+        Colname.grid_area,
+        f.col(grid_loss_responsible_energy_supplier).alias(Colname.energy_supplier_id),
+        Colname.time_window,
+        Colname.sum_quantity,
+        Colname.qualities,
+        Colname.metering_point_type,
+    )
+
+    df = (
+        result_df.join(
+            joined_grid_loss_result_and_responsible,
+            [Colname.time_window, Colname.grid_area, Colname.energy_supplier_id],
+            "outer",
+        )
+        .select(
+            Colname.grid_area,
+            result_df[Colname.balance_responsible_id],
+            f.coalesce(
+                result_df[Colname.energy_supplier_id],
+                joined_grid_loss_result_and_responsible[Colname.energy_supplier_id],
+            ).alias(Colname.energy_supplier_id),
+            Colname.time_window,
+            result_df[Colname.sum_quantity],
+            f.when(
+                result_df[Colname.qualities].isNull(),
+                joined_grid_loss_result_and_responsible[Colname.qualities],
+            )
+            .when(
+                joined_grid_loss_result_and_responsible[Colname.qualities].isNull(),
+                result_df[Colname.qualities],
+            )
+            .otherwise(
+                f.array_union(
+                    result_df[Colname.qualities], grid_loss_result_df[Colname.qualities]
+                )
+            )
+            .alias(Colname.qualities),
+            joined_grid_loss_result_and_responsible[Colname.sum_quantity].alias(
+                "grid_loss_sum_quantity"
+            ),
+        )
+        .na.fill(0, subset=["grid_loss_sum_quantity", Colname.sum_quantity])
+    )
+
+    result_df = df.withColumn(
+        adjusted_sum_quantity,
+        f.col(Colname.sum_quantity) + f.col("grid_loss_sum_quantity"),
+    )
+
+    result = result_df.select(
+        Colname.grid_area,
+        Colname.balance_responsible_id,
+        Colname.energy_supplier_id,
+        Colname.time_window,
+        f.col(adjusted_sum_quantity).alias(Colname.sum_quantity),
+        Colname.qualities,
+    ).orderBy(
+        Colname.grid_area,
+        Colname.balance_responsible_id,
+        Colname.energy_supplier_id,
+        Colname.time_window,
     )
 
     return EnergyResults(result)
