+++ source/databricks/calculation_engine/tests/calculation/energy/aggregations/grid_loss_aggregators/test_adjust_flex_consumption.py
@@ -23,13 +23,13 @@ from pyspark.sql.functions import col
 from pyspark.sql.types import (
     StructType,
     StringType,
+    DecimalType,
     TimestampType,
     BooleanType,
-    DecimalType,
     ArrayType,
 )
 
-from package.calculation.energy.aggregators.apply_grid_loss_adjustment import (
+from package.calculation.energy.aggregators.grid_loss_aggregators import (
     apply_grid_loss_adjustment,
 )
 from package.calculation.energy.energy_results import (
@@ -49,9 +49,9 @@ default_domain = "D1"
 default_responsible = "R1"
 default_supplier = "S1"
 default_sum_quantity = Decimal(1)
-default_negative_grid_loss = Decimal(3)
+default_positive_grid_loss = Decimal(3)
 default_aggregated_quality = QuantityQuality.ESTIMATED.value
-default_metering_point_type = MeteringPointType.PRODUCTION.value
+default_metering_point_type = MeteringPointType.CONSUMPTION.value
 
 date_time_formatting_string = "%Y-%m-%dT%H:%M:%S%z"
 default_time_window = {
@@ -67,9 +67,9 @@ default_valid_to = datetime.strptime(
 
 
 @pytest.fixture(scope="module")
-def sys_cor_schema() -> StructType:
+def grid_loss_sys_cor_schema() -> StructType:
     """
-    Input system correction data frame schema
+    Input grid loss system correction data frame schema
     """
     return (
         StructType()
@@ -80,12 +80,12 @@ def sys_cor_schema() -> StructType:
         .add(Colname.energy_supplier_id, StringType())
         .add(Colname.from_date, TimestampType())
         .add(Colname.to_date, TimestampType())
-        .add(Colname.is_negative_grid_loss_responsible, BooleanType())
+        .add(Colname.is_positive_grid_loss_responsible, BooleanType())
     )
 
 
 @pytest.fixture(scope="module")
-def hourly_production_result_row_factory(
+def flex_consumption_result_row_factory(
     spark: SparkSession,
 ) -> Callable[..., EnergyResults]:
     """
@@ -121,8 +121,8 @@ def hourly_production_result_row_factory(
 
 
 @pytest.fixture(scope="module")
-def negative_grid_loss_result_row_factory(
-    hourly_production_result_row_factory: Callable[..., EnergyResults],
+def positive_grid_loss_result_row_factory(
+    spark: SparkSession,
 ) -> Callable[..., EnergyResults]:
     """
     Factory to generate a single row of  data, with default parameters as specified above.
@@ -131,24 +131,32 @@ def negative_grid_loss_result_row_factory(
     def factory(
         domain: str = default_domain,
         time_window=None,
-        negative_grid_loss: Decimal = default_negative_grid_loss,
+        positive_grid_loss: Decimal = default_positive_grid_loss,
         aggregated_quality: str = default_aggregated_quality,
     ) -> EnergyResults:
         if time_window is None:
             time_window = default_time_window
-        return hourly_production_result_row_factory(
-            domain=domain,
-            time_window=time_window,
-            sum_quantity=negative_grid_loss,
-            aggregated_quality=aggregated_quality,
+        pandas_df = pd.DataFrame(
+            {
+                Colname.grid_area: [domain],
+                Colname.to_grid_area: [None],
+                Colname.from_grid_area: [None],
+                Colname.balance_responsible_id: ["balance_responsible_id"],
+                Colname.energy_supplier_id: ["energy_supplier_id"],
+                Colname.time_window: [time_window],
+                Colname.sum_quantity: [positive_grid_loss],
+                Colname.qualities: [[aggregated_quality]],
+            }
         )
+        df = spark.createDataFrame(pandas_df, schema=energy_results_schema)
+        return EnergyResults(df)
 
     return factory
 
 
 @pytest.fixture(scope="module")
-def sys_cor_row_factory(
-    spark: SparkSession, sys_cor_schema: StructType
+def grid_loss_sys_cor_row_factory(
+    spark: SparkSession, grid_loss_sys_cor_schema: StructType
 ) -> Callable[..., GridLossResponsible]:
     """
     Factory to generate a single row of  data, with default parameters as specified above.
@@ -162,44 +170,45 @@ def sys_cor_row_factory(
         supplier: str = default_supplier,
         valid_from: datetime = default_valid_from,
         valid_to: datetime = default_valid_to,
-        is_negative_grid_loss_responsible: bool = True,
+        is_positive_grid_loss_responsible: bool = True,
     ) -> GridLossResponsible:
         pandas_df = pd.DataFrame(
             {
-                Colname.metering_point_id: [metering_point_id],
+                metering_point_id: [metering_point_id],
                 Colname.grid_area: [domain],
                 Colname.metering_point_type: [metering_point_type],
                 Colname.balance_responsible_id: [responsible],
                 Colname.energy_supplier_id: [supplier],
                 Colname.from_date: [valid_from],
                 Colname.to_date: [valid_to],
-                Colname.is_negative_grid_loss_responsible: [
-                    is_negative_grid_loss_responsible
+                Colname.is_positive_grid_loss_responsible: [
+                    is_positive_grid_loss_responsible
                 ],
             }
         )
-        df = spark.createDataFrame(pandas_df, schema=sys_cor_schema)
+        df = spark.createDataFrame(pandas_df, schema=grid_loss_sys_cor_schema)
+
         return GridLossResponsible(df)
 
     return factory
 
 
-def test_grid_area_negative_grid_loss_is_added_to_grid_loss_responsible_energy_supplier(
-    hourly_production_result_row_factory: Callable[..., EnergyResults],
-    negative_grid_loss_result_row_factory: Callable[..., EnergyResults],
-    sys_cor_row_factory: Callable[..., GridLossResponsible],
+def test_grid_area_grid_loss_is_added_to_grid_loss_energy_responsible(
+    flex_consumption_result_row_factory: Callable[..., EnergyResults],
+    positive_grid_loss_result_row_factory: Callable[..., EnergyResults],
+    grid_loss_sys_cor_row_factory: Callable[..., GridLossResponsible],
 ) -> None:
     # Arrange
-    production = hourly_production_result_row_factory(supplier="A")
-    negative_grid_loss = negative_grid_loss_result_row_factory()
-    grid_loss_sys_cor_master_data = sys_cor_row_factory(supplier="A")
+    flex_consumption = flex_consumption_result_row_factory(supplier="A")
+    positive_grid_loss = positive_grid_loss_result_row_factory()
+    grid_loss_sys_cor_master_data = grid_loss_sys_cor_row_factory(supplier="A")
 
     # Act
     actual = apply_grid_loss_adjustment(
-        production,
-        negative_grid_loss,
+        flex_consumption,
+        positive_grid_loss,
         grid_loss_sys_cor_master_data,
-        MeteringPointType.PRODUCTION,
+        MeteringPointType.CONSUMPTION,
     )
 
     # Assert
@@ -207,26 +216,26 @@ def test_grid_area_negative_grid_loss_is_added_to_grid_loss_responsible_energy_s
         actual.df.where(col(Colname.energy_supplier_id) == "A").collect()[0][
             Colname.sum_quantity
         ]
-        == default_negative_grid_loss + default_sum_quantity
+        == default_positive_grid_loss + default_sum_quantity
     )
 
 
 def test_grid_area_grid_loss_is_not_added_to_non_grid_loss_energy_responsible(
-    hourly_production_result_row_factory: Callable[..., EnergyResults],
-    negative_grid_loss_result_row_factory: Callable[..., EnergyResults],
-    sys_cor_row_factory: Callable[..., GridLossResponsible],
+    flex_consumption_result_row_factory: Callable[..., EnergyResults],
+    positive_grid_loss_result_row_factory: Callable[..., EnergyResults],
+    grid_loss_sys_cor_row_factory: Callable[..., GridLossResponsible],
 ) -> None:
     # Arrange
-    production = hourly_production_result_row_factory(supplier="A")
-    negative_grid_loss = negative_grid_loss_result_row_factory()
-    grid_loss_sys_cor_master_data = sys_cor_row_factory(supplier="B")
+    flex_consumption = flex_consumption_result_row_factory(supplier="A")
+    positive_grid_loss = positive_grid_loss_result_row_factory()
+    grid_loss_sys_cor_master_data = grid_loss_sys_cor_row_factory(supplier="B")
 
     # Act
     actual = apply_grid_loss_adjustment(
-        production,
-        negative_grid_loss,
+        flex_consumption,
+        positive_grid_loss,
         grid_loss_sys_cor_master_data,
-        MeteringPointType.PRODUCTION,
+        MeteringPointType.CONSUMPTION,
     )
 
     # Assert
@@ -239,25 +248,25 @@ def test_grid_area_grid_loss_is_not_added_to_non_grid_loss_energy_responsible(
 
 
 def test_result_dataframe_contains_same_number_of_results_with_same_energy_suppliers_as_flex_consumption_result_dataframe(
-    hourly_production_result_row_factory: Callable[..., EnergyResults],
-    negative_grid_loss_result_row_factory: Callable[..., EnergyResults],
-    sys_cor_row_factory: Callable[..., GridLossResponsible],
+    flex_consumption_result_row_factory: Callable[..., EnergyResults],
+    positive_grid_loss_result_row_factory: Callable[..., EnergyResults],
+    grid_loss_sys_cor_row_factory: Callable[..., GridLossResponsible],
 ) -> None:
     # Arrange
-    hp_row_1 = hourly_production_result_row_factory(supplier="A")
-    hp_row_2 = hourly_production_result_row_factory(supplier="B")
-    hp_row_3 = hourly_production_result_row_factory(supplier="C")
+    fc_row_1 = flex_consumption_result_row_factory(supplier="A")
+    fc_row_2 = flex_consumption_result_row_factory(supplier="B")
+    fc_row_3 = flex_consumption_result_row_factory(supplier="C")
 
-    production = EnergyResults(hp_row_1.df.union(hp_row_2.df).union(hp_row_3.df))
-    negative_grid_loss = negative_grid_loss_result_row_factory()
-    grid_loss_sys_cor_master_data = sys_cor_row_factory(supplier="C")
+    flex_consumption = EnergyResults(fc_row_1.df.union(fc_row_2.df).union(fc_row_3.df))
+    positive_grid_loss = positive_grid_loss_result_row_factory()
+    grid_loss_sys_cor_master_data = grid_loss_sys_cor_row_factory(supplier="C")
 
     # Act
     actual = apply_grid_loss_adjustment(
-        production,
-        negative_grid_loss,
+        flex_consumption,
+        positive_grid_loss,
         grid_loss_sys_cor_master_data,
-        MeteringPointType.PRODUCTION,
+        MeteringPointType.CONSUMPTION,
     )
 
     # Assert
@@ -268,10 +277,10 @@ def test_result_dataframe_contains_same_number_of_results_with_same_energy_suppl
     assert actual_collect[2][Colname.energy_supplier_id] == "C"
 
 
-# def test_correct_negative_grid_loss_entry_is_used_to_determine_energy_responsible_for_the_given_time_window_from_hourly_production_result_dataframe(
-#     hourly_production_result_row_factory: Callable[..., EnergyResults],
-#     negative_grid_loss_result_row_factory: Callable[..., EnergyResults],
-#     sys_cor_row_factory: Callable[..., GridLossResponsible],
+# def test_correct_grid_loss_entry_is_used_to_determine_energy_responsible_for_the_given_time_window_from_flex_consumption_result_dataframe(
+#     flex_consumption_result_row_factory: Callable[..., EnergyResults],
+#     positive_grid_loss_result_row_factory: Callable[..., EnergyResults],
+#     grid_loss_sys_cor_row_factory: Callable[..., GridLossResponsible],
 # ) -> None:
 #     # Arrange
 #     time_window_1 = {
@@ -287,56 +296,56 @@ def test_result_dataframe_contains_same_number_of_results_with_same_energy_suppl
 #         Colname.end: datetime(2020, 1, 1, 3, 0),
 #     }
 #
-#     hp_row_1 = hourly_production_result_row_factory(
+#     fc_row_1 = flex_consumption_result_row_factory(
 #         supplier="A", time_window=time_window_1
 #     )
-#     hp_row_2 = hourly_production_result_row_factory(
+#     fc_row_2 = flex_consumption_result_row_factory(
 #         supplier="B", time_window=time_window_2
 #     )
-#     hp_row_3 = hourly_production_result_row_factory(
+#     fc_row_3 = flex_consumption_result_row_factory(
 #         supplier="B", time_window=time_window_3
 #     )
 #
-#     production = EnergyResults(hp_row_1.df.union(hp_row_2.df).union(hp_row_3.df))
+#     flex_consumption = EnergyResults(fc_row_1.df.union(fc_row_2.df).union(fc_row_3.df))
 #
-#     gasc_result_1 = Decimal(1)
-#     gasc_result_2 = Decimal(2)
-#     gasc_result_3 = Decimal(3)
+#     gagl_result_1 = Decimal(1)
+#     gagl_result_2 = Decimal(2)
+#     gagl_result_3 = Decimal(3)
 #
-#     gasc_row_1 = negative_grid_loss_result_row_factory(
-#         time_window=time_window_1, negative_grid_loss=gasc_result_1
+#     gagl_row_1 = positive_grid_loss_result_row_factory(
+#         time_window=time_window_1, positive_grid_loss=gagl_result_1
 #     )
-#     gasc_row_2 = negative_grid_loss_result_row_factory(
-#         time_window=time_window_2, negative_grid_loss=gasc_result_2
+#     gagl_row_2 = positive_grid_loss_result_row_factory(
+#         time_window=time_window_2, positive_grid_loss=gagl_result_2
 #     )
-#     gasc_row_3 = negative_grid_loss_result_row_factory(
-#         time_window=time_window_3, negative_grid_loss=gasc_result_3
-#     )
-#
-#     negative_grid_loss = EnergyResults(
-#         gasc_row_1.df.union(gasc_row_2.df).union(gasc_row_3.df)
+#     gagl_row_3 = positive_grid_loss_result_row_factory(
+#         time_window=time_window_3, positive_grid_loss=gagl_result_3
 #     )
 #
-#     sc_row_1 = sys_cor_row_factory(
+#     positive_grid_loss = EnergyResults(
+#         gagl_row_1.df.union(gagl_row_2.df).union(gagl_row_3.df)
+#     )
+#
+#     glsc_row_1 = grid_loss_sys_cor_row_factory(
 #         supplier="A", valid_from=time_window_1["start"], valid_to=time_window_1["end"]
 #     )
-#     sc_row_2 = sys_cor_row_factory(
+#     glsc_row_2 = grid_loss_sys_cor_row_factory(
 #         supplier="C", valid_from=time_window_2["start"], valid_to=time_window_2["end"]
 #     )
-#     sc_row_3 = sys_cor_row_factory(
+#     glsc_row_3 = grid_loss_sys_cor_row_factory(
 #         supplier="B", valid_from=time_window_3["start"], valid_to=None
 #     )
 #
 #     grid_loss_sys_cor_master_data = GridLossResponsible(
-#         sc_row_1.df.union(sc_row_2.df).union(sc_row_3.df)
+#         glsc_row_1.df.union(glsc_row_2.df).union(glsc_row_3.df)
 #     )
 #
 #     # Act
 #     actual = apply_grid_loss_adjustment(
-#         production,
-#         negative_grid_loss,
+#         flex_consumption,
+#         positive_grid_loss,
 #         grid_loss_sys_cor_master_data,
-#         MeteringPointType.PRODUCTION,
+#         MeteringPointType.CONSUMPTION,
 #     )
 #
 #     # Assert
@@ -345,7 +354,7 @@ def test_result_dataframe_contains_same_number_of_results_with_same_energy_suppl
 #         actual.df.where(col(Colname.energy_supplier_id) == "A")
 #         .filter(col(f"{Colname.time_window_start}") == time_window_1["start"])
 #         .collect()[0][Colname.sum_quantity]
-#         == default_sum_quantity + gasc_result_1
+#         == default_sum_quantity + gagl_result_1
 #     )
 #     assert (
 #         actual.df.where(col(Colname.energy_supplier_id) == "B")
@@ -357,5 +366,5 @@ def test_result_dataframe_contains_same_number_of_results_with_same_energy_suppl
 #         actual.df.where(col(Colname.energy_supplier_id) == "B")
 #         .filter(col(f"{Colname.time_window_start}") == time_window_3["start"])
 #         .collect()[0][Colname.sum_quantity]
-#         == default_sum_quantity + gasc_result_3
+#         == default_sum_quantity + gagl_result_3
 #     )
